requirements.txt
streamlit
pandas
plotly
tabula-py
PyPDF2
openpyxl
numpy
